# public/openapi.yaml
openapi: 3.0.3
info:
  title: Finanzas App API
  description: |
    API REST de la aplicación de finanzas personales (Next.js App Router + Prisma + PostgreSQL).

    ## Autenticación
    La mayoría de los endpoints requieren un **access token JWT** en el header
    `Authorization: Bearer <token>`. El token se obtiene en `/auth/login` o `/auth/register`.
    El **refresh token** se entrega como cookie `HttpOnly` (`refresh_token`) y se usa en `/auth/refresh`.

    ## Convenciones
    - Montos monetarios: número decimal (la API los serializa a veces como string desde PostgreSQL).
    - Fechas: ISO 8601 (`YYYY-MM-DDTHH:mm:ss.sssZ`).
    - Errores de validación → `422` con `{ error, details }`.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login correcto
          headers:
            Set-Cookie: { schema: { type: string }, description: refresh_token (HttpOnly) }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { description: Credenciales inválidas }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar access token
      description: Usa la cookie `refresh_token`. No requiere Authorization header.
      security: []
      responses:
        '200':
          description: Nuevo access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
        '401': { description: Refresh token ausente o inválido }
...
  /accounts/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [Accounts]
      summary: Detalle de una cuenta (incluye MSI activos)
      responses:
        '200': { description: Cuenta, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Accounts]
      summary: Actualizar cuenta
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountUpdate' }
      responses:
        '200': { description: Cuenta actualizada, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Accounts]
      summary: Eliminar cuenta (soft delete)
      responses:
        '200': { description: Cuenta eliminada }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────── TRANSACTIONS ─────────────────────
...
  /categories:
    get:
      tags: [Categories]
      summary: Categorías agrupadas por sección
      description: 'Devuelve un objeto `{ "Hogar": [...], "Ocio": [...] }` ordenado por sección.'
      responses:
        '200':
          description: Categorías agrupadas
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items: { $ref: '#/components/schemas/Category' }
    post:
      tags: [Categories]
      summary: Crear categoría de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name:  { type: string, maxLength: 60 }
                type:  { type: string, enum: [INCOME, EXPENSE] }
                color: { type: string, example: '#6366f1' }
                icon:  { type: string }
                group: { type: string, default: Otros }
      responses:
        '201': { description: Creada }
        '409': { description: Ya existe }
        '422': { $ref: '#/components/responses/ValidationError' }
  /categories/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    delete:
      tags: [Categories]
      summary: Eliminar categoría de usuario
      responses:
        '200': { description: Eliminada }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────────── MSI ──────────────────────────
...
  /analisis/gastos:
    get:
      tags: [Análisis]
      summary: Gastos agregados por categoría
      description: |
        Alimenta el donut de la pantalla Análisis (leyenda + total al centro).
        Reglas (idénticas a `/movimientos` para que el total == suma de las filas
        de un slice):
        - Sólo `type='EXPENSE'`.
        - Excluye pagos de tarjeta (`notes='TDC_PAYMENT'`) — null-safe.
        - **Incluye** filas MSI (`notes='MSI'`) a monto completo (llevan categoría real).
        - Ignora transacciones de cuentas inactivas (`account.isActive=true`).
        - Filas sin categoría se agrupan en un bucket sintético "Sin categoría".
      parameters:
        - { name: rango, in: query, schema: { type: string, enum: [semana, mes, semestre, anio], default: mes }, description: 'Ventana terminada en "now": semana=7d, mes=1m, semestre=6m, anio=1a' }
      responses:
        '200':
          description: Agregado por categoría
          content:
            application/json:
              schema:
                type: object
                properties:
                  rango: { type: string, enum: [semana, mes, semestre, anio] }
                  from:  { type: string, format: date-time }
                  to:    { type: string, format: date-time }
                  total: { type: number, description: Total del periodo }
                  categorias:
                    type: array
                    description: Ordenado desc por monto
                    items:
                      type: object
                      properties:
                        categoriaId: { type: string, nullable: true, description: 'null = Sin categoría' }
                        nombre:      { type: string }
                        color:       { type: string }
                        icon:        { type: string }
                        monto:       { type: number }
                        porcentaje:  { type: number, description: '% del total (1 decimal)' }
                        count:       { type: integer, description: Nº de movimientos }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access token JWT obtenido en /auth/login o /auth/register'
  responses:
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Datos inválidos
      content:
        application/json:
          schema:
            type: object
            properties:
              error:   { type: string, example: Datos inválidos }
              details: { type: object, additionalProperties: true }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        code:  { type: string }
    User:
      type: object
      properties:
        id:    { type: string }
        email: { type: string, format: email }
        name:  { type: string }
        plan:  { type: string, enum: [FREE, PREMIUM] }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        user:        { $ref: '#/components/schemas/User' }
    Account:
      type: object
      properties:
        id:             { type: string }
        name:           { type: string }
        type:           { type: string, enum: [CHECKING, SAVINGS, DEBIT, CREDIT, CASH, INVESTMENT, VOUCHER, FINTECH, DIGITAL_WALLET] }
        bank:           { type: string, nullable: true }
        color:          { type: string }
        icon:           { type: string, nullable: true }
        isActive:       { type: boolean }
        currentBalance: { type: number, description: 'TDC: deuda al corte + compras post-corte' }
        cutDebt:        { type: number, nullable: true, description: 'Deuda al corte del ciclo actual (TDC)' }
        creditLimit:    { type: number, nullable: true }
        cutDay:         { type: integer, nullable: true }
        paymentDueDay:  { type: integer, nullable: true }
        minimumPayment: { type: number, nullable: true }
        interestRate:   { type: number, nullable: true }
        activeInstallments: { type: integer, description: 'MSI activos vinculados' }
        totalMonthlyMSI:    { type: number, description: 'Suma de mensualidades MSI' }
    AccountCreate:
      type: object
      required: [name, type]
      properties:
        name:           { type: string, maxLength: 100 }
        type:           { type: string, enum: [CHECKING, SAVINGS, DEBIT, CREDIT, CASH, INVESTMENT] }
        bank:           { type: string }
        lastFourDigits: { type: string, pattern: '^\d{4}$' }
        color:          { type: string, example: '#6366f1' }
        icon:           { type: string }
        notes:          { type: string }
        currentBalance: { type: number, description: 'Saldo inicial (cuentas no-crédito)' }
        cutDebt:        { type: number, description: 'Deuda al corte inicial (sólo CREDIT)' }
        creditLimit:    { type: number }
        cutDay:         { type: integer, minimum: 1, maximum: 31 }
        paymentDueDay:  { type: integer, minimum: 1, maximum: 31 }
        minimumPayment: { type: number }
        interestRate:   { type: number, minimum: 0, maximum: 5 }
    AccountUpdate:
      allOf:
        - $ref: '#/components/schemas/AccountCreate'
        - type: object
          properties:
            isActive: { type: boolean }
...
    Category:
      type: object
      properties:
        id:       { type: string }
        name:     { type: string }
        type:     { type: string, enum: [INCOME, EXPENSE] }
        color:    { type: string }
        icon:     { type: string, nullable: true }
        group:    { type: string, example: Hogar }
        isSystem: { type: boolean }
...